feat: 알림 삭제에 대해 confirm 메세지 추가

+++ frontend/src/components/SubscriptionButton/useSubscriptionButton.ts
@@ -17,7 +17,9 @@ export const useSubscriptionButton = ({
   }, [isSubscribed]);
 
   const handleClickDeleteAllNotifications = () => {
-    deleteAllNotifications();
+    if (window.confirm('알림을 모두 삭제하시겠습니까?')) {
+      deleteAllNotifications();
+    }
   };
 
   const isAbleSubscribe =
